backend/requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
pandas
psycopg2-binary==2.9.1
sqlalchemy
slack-sdk==3.12.0
slackeventsapi
python-dotenv==1.0.0
openai==1.40.0
alembic
black
isort
flake8
apscheduler
python-multipart
redis  # 追記：Redisクライアントライブラリを追加
stripe
firebase-admin
functions_framework